@@ -3,8 +3,8 @@
     public interface IErrorDissipator
     {
         public void CreateEmptyMap(int width, int height);
-        public PixelError GetPixelError(int x, int y);
-        public void DissipateError(int x, int y, PixelError newColor);
+        public void GetPixelError(int x, int y, out PixelError error);
+        public void DissipateError(int x, int y, in PixelError newColor);
         public void ShiftLines();
     }
 }
